requirements.txt
pandas
numpy
plotly
requests
httpx
python-dotenv
sqlalchemy
psycopg2-binary
yfinance
postmarker

# FastHTML + streaming chat stack (replaces streamlit)
python-fasthtml
uvicorn[standard]

# LLM / agent
openai
langchain-core
langchain-openai
langgraph
